OPENMEETINGS-300 fixed
git-svn-id: https://svn.apache.org/repos/asf/incubator/openmeetings/trunk/singlewebapp@1344575 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/openmeetings/app/templates/FeedbackTemplate.java
@@ -34,15 +34,14 @@ public class FeedbackTemplate extends VelocityLoader {
 			FeedbackTemplate.class, OpenmeetingsVariables.webAppRootKey);
 
 	public String getFeedBackTemplate(String username, String email,
-			String message, Integer default_lang_id) {
+			String message, long default_lang_id) {
 		try {
 
 			super.init();
 
 			// TODO: Finish Feedback - Template
-			// Fieldlanguagesvalues fValue =
-			// fieldmanagment.getFieldByIdAndLanguage(new Long(499), new
-			// Long(default_lang_id));
+			// String fValue =
+			// fieldmanagment.getString(499L, default_lang_id);
 
 			/* lets make a Context and put data into it */
